requirements.txt
pytest
PyContracts